requirements.txt
# Core dependencies
PyQt6>=6.6.0
numpy>=1.20.0
typing-extensions>=4.5.0
huggingface-hub>=0.20.0

# Model backends
# For GGUF files (quantized models)
llama-cpp-python>=0.3.0

# For original models (recommended for BF16/F16)
transformers>=4.51.3
torch>=2.0.0
accelerate>=1.0.0
